payrollcycle list order fix

+++ src/main/java/com/witcurve/service/util/PayrollCycleComparator.java
@@ -7,7 +7,7 @@ import java.util.Comparator;
 public class PayrollCycleComparator implements Comparator<PayrollCycleDTO> {
     @Override
     public int compare(PayrollCycleDTO o1, PayrollCycleDTO o2) {
-        if (o1.getYear() == o2.getYear()) {
+        if (o1.getYear().intValue() == o2.getYear().intValue()) {
             if (o1.getMonth().getValue() == o2.getMonth().getValue()) {
                 return 0;
             } else if (o1.getMonth().getValue() > o2.getMonth().getValue()) {
@@ -15,7 +15,7 @@ public class PayrollCycleComparator implements Comparator<PayrollCycleDTO> {
             } else {
                 return -1;
             }
-        } else if (o1.getYear() > o2.getYear()) {
+        } else if (o1.getYear().intValue() > o2.getYear().intValue()) {
             return 1;
         } else {
             return -1;
